+++ CHANGELOG.md
@@ -7,7 +7,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
-TODO: add at least one Added, Changed, Deprecated, Removed, Fixed or Security section
+### Changed
+
+- Use cmem-client instead of cmempy for all Corporate Memory access
+
+### Removed
+
+- Remove cmem-cmempy dependency and `setup_cmempy_user_access` calls
 
 ## [2.2.7] 2026-07-22